requirements.txt
numpy
openai
python-dotenv
tiktoken
flask
opencv-python
uuid
PyPDF2